Added xml specification and parsing as well as allowed whitelist.
XML specification of mappings would have required an eval() of a string. To
avoid this security disaster, any desired methods must be specifically "allowed"
and passed to the mapping tool. This brings the methods into context and acts
as a whitelist of acceptable methods. If the code is later refactored to send
commands to clients/servers or similar, the whitelist can remain but the passing
of the methods into scope will no longer be needed.